+++ NEWS.md
@@ -1,3 +1,9 @@
+# restez (in development)
+
+## Bug fixes
+
+* Check internet connection in mainland China
+
 # restez 1.0.0
 
 ## Post-review version of `restez` released.